config: Reword bltouch description in example-extras.cfg
The pin_up_touch_mode_reports_triggered should be set to True on all
genuine BLTouch.

Expand the description for stow_on_each_sample and state the default.

Reword the set_output_mode option to be more similar to other
descriptions.

It is no longer necessary to warn about clones being not supported.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>