config: Updates to TAZ6 single extruder config (#3750)
The Lulzbot TAZ6 single extruder configuration had an XY position for
z homing that did not match the factory firmware, and was off center
enough to cause the extruder to crash into the home button trim
sometimes when homing. This commit changes the printer config so
that the XY position for Z homing matches the original Marlin firmware
for this printer.

The Lulzbot TAZ6 single extruder configuration had an bed_tilt points
that did not match the factory firmware, and were off enough that the
nozzle (used for probing on this printer) could slip off the edge of
the probe point or miss it entirely. This commit changes the printer
config so that the coordinates for bed_tilt calibration match the
original Marlin firmware for this printer.

For the Lulzbot TAZ6 single extruder config, use 'gear_ratio' and
'rotation_distance' to configure the extruder.

The rotation distance is based on effective circumference of the hobbed
bolt based on the original Marlin firmware for the TAZ6 single-extruder,
and the gear ratio was determined by simply counting the number of
teeth on the pinion and spur gears.

Signed-off-by: Steven T. Snyder <github@steventsnyder.com>